+++ src/core/quizHandler/preStart.js
@@ -5,13 +5,13 @@ import loadTemplate from "../../templates/load.js";
 import preStartTemplate from "../../templates/prestart.html?raw";
 
 export default async function handlePreStartState({ startBtn, storedQuiz }) {
-    window.location.hash = "";
+    // window.location.hash = "";
 
     startBtn.disabled = true;
 
     setEduQuickStatus("Fetched all answers.");
     setEduQuickTitle("Before you begin");
-    setEduQuickContent(await loadTemplate(preStartTemplate));
+    setEduQuickContent(loadTemplate(preStartTemplate));
 
     const original = startBtn.textContent.trim();
     let t = 10;
