--- src/ui/layout.js
@@ -1,9 +1,0 @@
-export function applyLayout() {
-    document.body.style.marginRight = "25%";
-    document.body.style.marginBottom = "8%";
-}
-
-export function resetLayout() {
-    document.body.style.marginRight = "";
-    document.body.style.marginBottom = "";
-}
