+++ src/ui/sidebar.js
@@ -5,7 +5,7 @@ export function createSidebar() {
     sidebar.id = "eduquick-sidebar";
     sidebar.className =
         "bg-purple border-purple-30 text-white flex flex-col p-4 pb-0";
-
+    // TODO: use tailwindcss or css, not both
     Object.assign(sidebar.style, {
         position: "fixed",
         top: "0",
@@ -16,29 +16,30 @@ export function createSidebar() {
         borderLeftWidth: "3px",
     });
 
-    const title = document.createElement("div");
-    title.id = "eduquick-sidebar-title";
-    title.className = "border-b-3 border-purple-30 font-bold text-3xl mb-4";
+    const sidebarTitle = document.createElement("div");
+    sidebarTitle.id = "eduquick-sidebar-title";
+    sidebarTitle.className =
+        "border-b-3 border-purple-30 font-bold text-3xl mb-4";
+    sidebar.appendChild(sidebarTitle);
 
-    const content = document.createElement("div");
-    content.id = "eduquick-sidebar-content";
-    content.className = "flex-1";
+    const sidebarContent = document.createElement("div");
+    sidebarContent.id = "eduquick-sidebar-content";
+    sidebarContent.className = "flex-1";
+    sidebar.appendChild(sidebarContent);
 
-    sidebar.appendChild(title);
-    sidebar.appendChild(content);
     document.body.appendChild(sidebar);
 }
 
-export function setEduQuickTitle(text) {
-    const titleEl = document.getElementById("eduquick-sidebar-title");
-    if (titleEl) {
-        titleEl.textContent = text;
+export function setEduQuickTitle(title) {
+    const sidebarTitle = document.getElementById("eduquick-sidebar-title");
+    if (sidebarTitle) {
+        sidebarTitle.textContent = title;
     }
 }
 
 export function setEduQuickContent(content) {
-    const contentEl = document.getElementById("eduquick-sidebar-content");
-    if (contentEl) {
-        contentEl.innerHTML = content;
+    const sidebarContent = document.getElementById("eduquick-sidebar-content");
+    if (sidebarContent) {
+        sidebarContent.innerHTML = content;
     }
 }
